requirements.txt
numpy~=2.3.2
pandas~=2.3.1
torch~=2.7.1
pillow~=11.3.0